Rename Table Tennis game entry to Ping Pong and fix route wiring
